% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{clear}
\alias{clear}
\title{Drain RethinkDB connection}
\usage{
clear(connection)
}
\arguments{
\item{connection}{structure with rethinkdb connection details}
}
\description{
Drain RethinkDB connection
}
