% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{clone_connection}
\alias{clone_connection}
\title{Clone connection}
\usage{
clone_connection(connection)
}
\arguments{
\item{connection}{structure with rethinkdb connection details}
}
\value{
copy of connection structure
}
\description{
Clone connection
}
\examples{
\dontrun{
cn <- connect()
cn_copy <- clone_connection(cn)
}
}
