% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{collection}
\alias{collection}
\title{Collect table entries}
\usage{
collection(collection_name, connection, column_names = character(),
  post_process = I)
}
\arguments{
\item{collection_name}{A name of collection.}

\item{connection}{Connection object.}

\item{column_names}{Names of specific columns.}

\item{post_process}{Post process function called before query is run.}
}
\value{
Reactive value with connection, collection_name and collection
}
\description{
This collects entries from specific collection and assigns it to
a reactive value.
}
\examples{
\dontrun{
# Set up a connection to DEFAULT_DB
cn <- connect()
tvs <- collection("tv_shows", cn)
# Take a look at gathered entries
isolate(tvs$collection)

# Using post_process you might use specific rethinkDB operations
# in rethinker query format before query is run (like filter, orderBy).
# E.g., to gather only tv shows with less than 100 episodes use:
tvs <- collection("tv_shows", cn,
            post_process = function(q) q$filter(function(x) x$bracket("episodes")$lt(100)))
}
}
