% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Connect with RethinkDB}
\usage{
connect(host = "localhost", port = "28015", db_name = DEFAULT_DB,
  config_file = DEFAULT_CONFIG_NAME)
}
\arguments{
\item{host}{character with host address (default `localhost`)}

\item{port}{character with port number (default `28015`)}

\item{db_name}{character with databse name (default `DEFAULT_DB`)}

\item{config_file}{character with rethink configuration file name
(default `DEFAULT_CONFIG_NAME`)}
}
\value{
connection structure
}
\description{
This function reads config file (if exists) or from `host` and `port`
from parameters. Next, it opens a connection and creates database `db_name`
(only if  already does not exists).
}
\examples{
\dontrun{
cn <- connect()
}
}
