% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cursor_to_tibble}
\alias{cursor_to_tibble}
\title{Make tibble from cursor}
\usage{
cursor_to_tibble(cursor, column_names)
}
\arguments{
\item{cursor}{rethinkDB cursor}

\item{column_names}{character vector with column names}
}
\value{
either empty tibble or tibble with data
}
\description{
Receives data in tibble format from rethinkDB cursor.
}
