% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{DEFAULT_CONFIG_NAME}
\alias{DEFAULT_CONFIG_NAME}
\title{Path to rethinkdb configuration file.}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_CONFIG_NAME
}
\description{
Path to rethinkdb configuration file.
}
\keyword{datasets}
