% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{DEFAULT_DB}
\alias{DEFAULT_DB}
\title{Default database name.}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_DB
}
\description{
Default database name.
}
\keyword{datasets}
