% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete}
\alias{delete}
\title{Delete element from collection}
\usage{
delete(collection, element_id)
}
\arguments{
\item{collection}{reactive value with collection data}

\item{element_id}{character with element id}
}
\value{
result status with numer of elements: deleted, errored,
inserted, replaced, skipped, unchanged
}
\description{
Delete element from collection
}
