% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_handle}
\alias{get_table_handle}
\title{Get table handle}
\usage{
get_table_handle(db_name, collection_name)
}
\arguments{
\item{db_name}{character with database name}

\item{collection_name}{character with collection name}
}
\value{
reql object
}
\description{
Get table handle
}
\examples{
\dontrun{
 get_table_handle("shiny", "cars")
}
}
