% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{insert}
\alias{insert}
\title{Insert element to collection}
\usage{
insert(collection, element, ...)
}
\arguments{
\item{collection}{reactive value with collection data}

\item{element}{list with content of element}

\item{...}{other params to insert}
}
\value{
result status with numer of elements: deleted, errored,
inserted, replaced, skipped, unchanged
}
\description{
Insert element to collection
}
