% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{make_sure_db_exists}
\alias{make_sure_db_exists}
\title{Make sure that rethink database exists}
\usage{
make_sure_db_exists(connection, db_name)
}
\arguments{
\item{connection}{structure with rethinkdb connection details}

\item{db_name}{character with database name}
}
\description{
This function checks whether database `db_name` exists and if not
it creates one.
}
\examples{
\dontrun{
cn <- connect()
make_sure_db_exists(cn, "temp_db")
}
}
