% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{make_sure_table_exists}
\alias{make_sure_table_exists}
\title{Make sure that rethinkdb table exists}
\usage{
make_sure_table_exists(connection, table_name)
}
\arguments{
\item{connection}{structure with rethinkdb connection details}

\item{table_name}{character with table name}
}
\description{
This function checks whether table `table_name` exists in databse `db_name`
and if not it creates one.
}
